% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{calcBescape}
\alias{calcBescape}
\title{Calculate the fishing mortality required to get the assigned TAC}
\usage{
calcBescape(Bpa, df.tmb, N_current, Fsel, Fcap = 2)
}
\arguments{
\item{df.tmb}{SMS input list}

\item{N_current}{numbers at age at the beginning of the year}

\item{TAC}{TAC were are trying to achieve}

\item{bios}{list of bioparams}

\item{findTAC}{Look for TAC or SSB}
}
\description{
Calculate the fishing mortality required to get the assigned TAC
}
