% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{calcFTAC}
\alias{calcFTAC}
\title{Calculate the fishing mortality required to get the assigned TAC}
\usage{
calcFTAC(TAC, df.tmb, bios, Ncurrent, findTAC = TRUE)
}
\arguments{
\item{TAC}{TAC were are trying to achieve}

\item{df.tmb}{SMS input list}

\item{bios}{list of bioparams}

\item{findTAC}{Look for TAC or SSB}

\item{N_current}{numbers at age at the beginning of the year}
}
\description{
Calculate the fishing mortality required to get the assigned TAC
}
