% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{calcTAC}
\alias{calcTAC}
\title{Forecast an sms model}
\usage{
calcTAC(
  df.tmb,
  sas,
  recruitment = "hockey",
  HCR = "Fmsy",
  avg_years = rep(3, 4)
)
}
\arguments{
\item{df.tmb}{sms dataframe}

\item{sas}{fitted sms model}

\item{recruitment}{hockey, long_mean or short_mean}

\item{HCR}{Fmsy or Bescape}

\item{avg_years}{years to average weca, M, west, and mat}

\item{method}{}
}
\description{
Forecast an sms model
}
