% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tmb_error.R
\name{check_tmb_error}
\alias{check_tmb_error}
\title{Check if parameter inputs has the correct format and dimensions to make sure TMB does not crash R}
\usage{
check_tmb_error(df.tmb)
}
\arguments{
\item{df.tmb}{}
}
\description{
Check if parameter inputs has the correct format and dimensions to make sure TMB does not crash R
}
\examples{
check_tmb_error(sandeel1r$df.tmb)

}
