% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{createForecastTable}
\alias{createForecastTable}
\title{Produce an ICES forecast table}
\usage{
createForecastTable(
  df.tmb,
  sas,
  TACold,
  HCR = "Bescape",
  avg_R = "mean",
  Bpa = NULL
)
}
\arguments{
\item{df.tmb}{sms model input data}

\item{sas}{fitted sms model}

\item{TACold}{last years tac}

\item{HCR}{Harvest control rule to use - options Bescape or Fmsy}

\item{avg_R}{years to average recruitment}

\item{Bpa}{SSB at Bpa}
}
\description{
Produce an ICES forecast table
}
