% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_matrix.R
\name{df_to_matrix}
\alias{df_to_matrix}
\title{Convert dataframes into matrices}
\usage{
df_to_matrix(df, season = 1:4)
}
\arguments{
\item{df}{data frame containing data}

\item{season}{number of quarters in model}
}
\value{
Returns the dataframe as a matrix for tmb
}
\description{
Convert dataframes into matrices
}
\examples{
df_to_matrix(df_lol, season = 1:4)
}
