% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast.sms}
\alias{forecast.sms}
\title{Simple one year forecast}
\usage{
forecast.sms(df.tmb, N_current, F0)
}
\arguments{
\item{F0}{}
}
\value{
returns a list with forecasted SSB and catch
}
\description{
Simple one year forecast
}
