% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_EM_parameters.R
\name{get_EM_parameters}
\alias{get_EM_parameters}
\title{Get parameters from an operating model}
\usage{
get_EM_parameters(df.OM, OM, df.tmb, obsSurvey = FALSE, obsCatch = FALSE)
}
\arguments{
\item{df.OM}{operating model input}

\item{OM}{smsR fitted operating model}

\item{df.tmb}{data input for assessment model used to condition OM}

\item{obsSurvey}{should observed or modeled survey be used}

\item{obsCatch}{should observed or modeled catch be used}
}
\value{
returns a list with smsR data input
}
\description{
Get parameters from an operating model
}
\examples{
get_EM_parameters(...)
}
