% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getBiomass}
\alias{getBiomass}
\title{Biomass at age}
\usage{
getBiomass(df.tmb, sas)
}
\arguments{
\item{df.tmb}{list of input parameters}

\item{sas}{fitted smsR model}
}
\value{
data frame containing the biomass of each age group.
low and high are in the 95\% confidence region.
SE is standard error of log Biomass
}
\description{
Retrieves a data frame containing the biomass per age from an `smsR` stock assessment
}
\seealso{
[getR()],[getYield()],[getF()],[getCatch()], [getSSB()], [getN()], [getFbar()], [getCatchN()], [getCatchCV()],
[getSurvey()]
}
