% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getCatchability}
\alias{getCatchability}
\title{Survey catchability}
\usage{
getCatchability(df.tmb, sas)
}
\arguments{
\item{df.tmb}{sms input parameters}

\item{sas}{fitted model object}
}
\value{
returns a data frame of survey catchability
}
\description{
Survey catchability
}
\details{
Exports a data frame of catchabilities from all the surveys in a fitted `smsR` model.

Note that the scale of the catchability can vary based on input data.
}
\examples{

print(getCatchability(df.tmb, sas))

}
