% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getCatchCV}
\alias{getCatchCV}
\title{Retrieve estimated parameters}
\usage{
getCatchCV(sas)
}
\arguments{
\item{sas}{stock assessment output from smsR}
}
\value{
data frame of estimated parameters and their SE
}
\description{
Retrieve estimated parameters
}
