% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataSandeel.R
\name{getDataSandeel}
\alias{getDataSandeel}
\title{Title}
\usage{
getDataSandeel(
  wd,
  maxage = NA,
  survey.age = list(),
  survey.years = list(),
  survey.names = NA,
  survey.quarter = NA,
  effort = 0,
  years,
  seasons
)
}
\arguments{
\item{wd}{working directory}

\item{maxage}{maxage}

\item{survey.age}{list showing ages in surveys}

\item{survey.years}{years ssurveys occur}

\item{survey.names}{names of survey}

\item{survey.quarter}{which quarters do surveys happen}

\item{effort}{set to true to read in effort data}

\item{years}{years to calculate}

\item{seasons}{seasons to calc}
}
\value{

}
\description{
Title
}
\examples{
x <- getDataSandeel # Not run
}
