% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataSMS.R
\name{getDataSMS}
\alias{getDataSMS}
\title{Read data from an SMS model}
\usage{
getDataSMS(
  wd,
  maxage = NA,
  survey.age = list(),
  survey.years = list(),
  survey.names = NA,
  survey.quarter = NA,
  effort = FALSE,
  years,
  seasons,
  scv.tv = 0
)
}
\arguments{
\item{wd}{working directory}

\item{maxage}{maxage}

\item{survey.age}{list showing ages in surveys}

\item{survey.years}{years ssurveys occur}

\item{survey.names}{names of survey}

\item{survey.quarter}{which quarters do surveys happen}

\item{effort}{set to true to read in effort data}

\item{years}{years to calculate}

\item{seasons}{seasons to calc}
}
\value{
Returns a list of data frames
}
\description{
Read data from an SMS model
}
\examples{
wd <- 'C:/testdir/'
x <- getDataSandeel(wd, maxage = 5, survey.age = c(1:2),
survey.years = 2004:2023, survey.quarter = 1, years = 1985:2024, seasons = 1:2) # Not run
}
