% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getEstimatedParms}
\alias{getEstimatedParms}
\title{Get fishin mortality at age from fitted model}
\usage{
getEstimatedParms(df.tmb, sas)
}
\arguments{
\item{df.tmb}{list of input parameters}

\item{sas}{fitted smsR model}
}
\value{
data frame of fishing mortality at age. Last year is based on the expected survival from terminal year. minSE and maxSE is the 95% confidence intervals
}
\description{
Get fishin mortality at age from fitted model
}
