% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getFbar}
\alias{getFbar}
\title{Title}
\usage{
getFbar(df.tmb, sas, Fbarage = NULL)
}
\arguments{
\item{df.tmb}{list of input parameters}

\item{sas}{smsR fitted model}

\item{Fbarage}{ages to calculate Fbar}
}
\description{
Title
}
