% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFmsy_OM.R
\name{getFmsy.OM}
\alias{getFmsy.OM}
\title{Estimate Fmsy from a SMS Operating Model}
\usage{
getFmsy.OM(
  OM,
  df,
  nyears = 50,
  recruitment = "hockey",
  plotMSY = FALSE,
  stochastic = FALSE
)
}
\arguments{
\item{OM}{An sms operating model from \code{"run.agebased.sms.op()"}}

\item{df}{A list containing operating model parameters.}

\item{nyears}{Integer. Number of years to simulate (default is 50).}

\item{recruitment}{Character. Recruitment model to use in the simulation. Currently only \code{"hockey"} is supported.}

\item{plotMSY}{Logical. If \code{TRUE}, a plot of the yield curve and Fmsy estimate is produced.}

\item{stochastic}{Logical. Should stochastic recruitment deviations be included in the simulation?}
}
\value{
A numeric value representing the fishing mortality rate that results in maximum sustainable yield (Fmsy).
}
\description{
Calculates the fishing mortality rate (\emph{F}) that produces maximum sustainable yield (Fmsy)
using an SMS operating model. This function simulates forward under a
specified recruitment assumption and identifies the \emph{F} that maximizes yield.
}
\examples{
\dontrun{
fmsy <- getFmsy.OM(sas = fitted_model, df = om_params, nyears = 100,
                   recruitment = "hockey", plotMSY = TRUE, stochastic = TRUE)
}
}
