% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getM_var}
\alias{getM_var}
\title{get the natural mortality at age from the assesment model}
\usage{
getM_var(df.tmb, sas, plotFig = FALSE)
}
\arguments{
\item{df.tmb}{smsR list of input parameters}

\item{plotFig}{plot the data (TRUE OR FALSE)}
}
\value{
returns a data frame of weight at aeg
}
\description{
get the natural mortality at age from the assesment model
}
\examples{

M <- getM(df.tmb, plotFig = TRUE) # Plot the natural mortality by age

}
