% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getMat}
\alias{getMat}
\title{get the maturity at age from the input data and plot it}
\usage{
getMat(df.tmb, plotFig = FALSE)
}
\arguments{
\item{df.tmb}{smsR list of input parameters}

\item{plotFig}{plot the data (TRUE OR FALSE)}
}
\value{
returns a data frame of weight at aeg
}
\description{
get the maturity at age from the input data and plot it
}
\examples{
getMat(df.tmb, plotFig = TRUE) # Plot maturity

}
