% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParms.R
\name{getMPS}
\alias{getMPS}
\title{Map smsR parameters}
\usage{
getMPS(df.tmb, parms, mapExtra = NA)
}
\arguments{
\item{df.tmb}{list of tmb input data}

\item{parms}{list of estimated parameters}

\item{mapExtra}{vector of parameters to map}
}
\value{
returns a list of parameters to be mapped for smsR
}
\description{
Retrieve a list of mapped parameters (not estimated) from an `smsR` dataset
}
\examples{

# Map the  CV of surveys to the parms value
MPS <- getMPS(df.tmb, parms, mapExtra = "SDsurvey")

}
\seealso{
\code{\link{getParms}}
}
