% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{getOM_FTAC}
\alias{getOM_FTAC}
\title{calc f required in OM to get TAC}
\usage{
getOM_FTAC(TAC, df.OM, Fcap = 10)
}
\arguments{
\item{TAC}{TAC to optimize for}

\item{df.OM}{list of OM parameters}

\item{Fcap}{Maximum F to test}
}
\description{
calc f required in OM to get TAC
}
