% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getR}
\alias{getR}
\title{Get a data frame of recruitment}
\usage{
getR(df.tmb, sas)
}
\arguments{
\item{df.tmb}{list of input parameters}

\item{sas}{fitted smsR model}
}
\value{
data frame of recruitment. Last year is based on the Stock recruitment relationship. minSE and maxSE is the 95% confidence intervals
}
\description{
Get a data frame of recruitment
}
