% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getResidCatch}
\alias{getResidCatch}
\title{Catch residuals}
\usage{
getResidCatch(df.tmb, sas)
}
\arguments{
\item{df.tmb}{list of input data for smsR object}

\item{sas}{fitted stock assessment model for smsR}
}
\value{
Returns a data frame of catch residuals
}
\description{
Get the absolute catch residuals for a fitted `smsR` model.
}
\details{
To plot the CV scaled catch residuals see \code{\link{plotDiagnostics}}
}
\examples{

getResidCatch(df.tmb, sas)
print(getResidCatch)

}
\seealso{
\code{\link{getResidSurvey}},\code{\link{plotDiagnostics}}
}
