% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getResidSurvey}
\alias{getResidSurvey}
\title{Survey residuals}
\usage{
getResidSurvey(df.tmb, sas)
}
\arguments{
\item{df.tmb}{list of input parameters}

\item{sas}{fitted smsR model}
}
\value{
data frame of absolute survey residuals
}
\description{
Get survey residuals from a fitted `smsR` model
}
\examples{
res <- getResidSurvey(df.tmb, sas)
print(res)

}
\seealso{
\code{\link{getResidCatch}},\code{\link{plotDiagnostics}}
}
