% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getSel}
\alias{getSel}
\title{Fishing selectivity}
\usage{
getSel(df.tmb, sas)
}
\arguments{
\item{df.tmb}{sms input parameters}

\item{sas}{fitted sms model}
}
\value{
returns a table of estimated fisheries selectivity
}
\description{
Retrieve the fishing selectivity from a fitted `smsR` object. The output from this function is scaled to the maximum selectivity per selectivity block.
}
\examples{
## ' Not Run
sel <- getSel(df.tmb, sas)

## End
}
