% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getSelectivity}
\alias{getSelectivity}
\title{Get the selectivity in the last year}
\usage{
getSelectivity(df.tmb, sas)
}
\arguments{
\item{df.tmb}{list of input data for an smsR model}

\item{sas}{fitted smsR model}
}
\value{
returns a table of most recent selectivity with uncertainty
}
\description{
Get the selectivity in the last year
}
\examples{

sel <- getSelectivity(df.tmb, sas)

}
