% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getSR}
\alias{getSR}
\title{Stock recruitment}
\usage{
getSR(df.tmb, sas, method = "mean")
}
\arguments{
\item{df.tmb}{data frame with input data}

\item{sas}{fitted smsR model}

\item{method}{mean or median recruitment}
}
\value{
returns a data frame with a stock recruitment relationship
}
\description{
Get the estimated stock recruitment relationship from a fitted `smsR` model
}
\examples{

SR <- getSR(df.tmb, sas)

}
