% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getSummaryCVs}
\alias{getSummaryCVs}
\title{Get the CVs of the summary information}
\usage{
getSummaryCVs(df.tmb, sas, verbose = TRUE)
}
\arguments{
\item{verbose}{}
}
\description{
Get the CVs of the summary information
}
