% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getSurvey}
\alias{getSurvey}
\title{Get the estimated survey}
\usage{
getSurvey(df.tmb, sas)
}
\arguments{
\item{df.tmb}{list of input data for smsR}

\item{sas}{fitted smsR assessment model}
}
\value{
returns the estimated survey as a data frame
}
\description{
Get the estimated survey
}
\examples{

survest <- getSurvey(df.tmb, sas)

}
