% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getSurveyCV}
\alias{getSurveyCV}
\title{Survey CV
Retrieve estimated survey CV from a fitted `smsR` model.}
\usage{
getSurveyCV(df.tmb, sas)
}
\arguments{
\item{sas}{stock assessment output from smsR}
}
\value{
data frame of estimated parameters and their SE
}
\description{
Survey CV
Retrieve estimated survey CV from a fitted `smsR` model.
}
