% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getSurveySD}
\alias{getSurveySD}
\title{Retrieve estimated survey SD from a fitted `smsR` model.}
\usage{
getSurveySD(df.tmb, sas)
}
\arguments{
\item{sas}{stock assessment output from smsR}
}
\value{
data frame of estimated parameters and their SE
}
\description{
Retrieve estimated survey SD from a fitted `smsR` model.
}
