% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{getTAC}
\alias{getTAC}
\title{Forecast an sms model}
\usage{
getTAC(
  df.tmb,
  sas,
  recruitment = "hockey",
  HCR = "Fmsy",
  avg_R = NULL,
  Bpa = NULL,
  Fcap = NULL
)
}
\arguments{
\item{df.tmb}{sms dataframe}

\item{sas}{fitted sms model}

\item{recruitment}{hockey, long_mean or short_mean}

\item{HCR}{Fmsy or Bescape}

\item{avg_R}{years to average R}

\item{Bpa}{SSB value for Bpa}

\item{Fcap}{Maximum possible fishing mortality}
}
\description{
Forecast an sms model
}
