% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getTSB}
\alias{getTSB}
\title{Total stock biomass}
\usage{
getTSB(df.tmb, sas)
}
\arguments{
\item{df.tmb}{list of input parameters}

\item{sas}{fitted smsR model}
}
\value{
data frame containing the biomass of each age group.
low and high are the 95\% confidence intervals.
SE is standard error of log TSB
}
\description{
Export the total stock biomass summed over ages in the first season from a fitted `smsR` stock assessment
}
\examples{

dat <- getTSB(df.tmb, sas)

}
\seealso{
\code{\link{getR}}, \code{\link{getSSB}}, \code{\link{getYield}},
  \code{\link{getF}}, \code{\link{getCatch}}, \code{\link{getBiomass}},
  \code{\link{getN}}, \code{\link{getFbar}}, \code{\link{getCatchN}},
  \code{\link{getCatchCV}}, \code{\link{getSurvey}}
}
