% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getWeight}
\alias{getWeight}
\title{get the weight at age from the input data and plot it}
\usage{
getWeight(df.tmb, WW = "weca", plotFig = FALSE)
}
\arguments{
\item{df.tmb}{smsR list of input parameters}

\item{WW}{weca or west}

\item{plotFig}{plot the data (TRUE OR FALSE)}
}
\value{
returns a data frame of weight at aeg
}
\description{
get the weight at age from the input data and plot it
}
\examples{
getWeight(df.tmb, WW = "west", plotFig = TRUE) # Plots the weight in the stock
}
