% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getYieldSeason}
\alias{getYieldSeason}
\title{Get the seasonal input yield}
\usage{
getYieldSeason(df.tmb)
}
\arguments{
\item{df.tmb}{a list of parameters for smsR model from \link{get_TMB_parameters()}}
}
\value{
a data frame of yield
}
\description{
Get the seasonal input yield
}
\examples{

Yield <- getYieldSeason(df.tmb) # Get observed catch at age as a data frame

}
