% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{logLik.sms}
\alias{logLik.sms}
\title{Extract the log likelihood of a sms model}
\usage{
\method{logLik}{sms}(object, ...)
}
\value{
object of class \code{logLik} with attributes
\item{val}{log likelihood}
\item{nobs,nall}{number of non NA observations initially supplied to TMB}
\item{df}{number of parameters}
}
\description{
Gets the log likelihood of a fitted `smsR` model
}
