% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mohns_rho.R
\name{mohns_rho}
\alias{mohns_rho}
\title{Calculate mohns rho.}
\usage{
mohns_rho(
  df.tmb,
  peels = 5,
  parms,
  mps,
  lower = -Inf,
  upper = Inf,
  Fbarage = c(1, 2),
  plotfigure = TRUE,
  limits = NA,
  dll = "sms"
)
}
\arguments{
\item{df.tmb}{list of tmb parameters}

\item{peels}{number of peels}

\item{parms}{parameters that are estimated}

\item{mps}{parameters that are mapped}

\item{lower}{lower boundary of parameters}

\item{upper}{upper boundary of parameters}

\item{Fbarage}{ages to calculate Fbar}

\item{plotfigure}{TRUE FALSE of whether to plot figure}

\item{limits}{limits on plot}

\item{dll}{which dll to use. Uses 'sms' as standard}
}
\description{
Calculate mohns rho.
}
\examples{
x <- mohns_rho(df.tmb, parms = parms, mps = mps, Fbarage = 3:6)
}
