% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{optTAC}
\alias{optTAC}
\title{Optimizer for finding the fishing mortality that leads to TAC}
\usage{
optTAC(data, par)
}
\arguments{
\item{par}{}
}
\description{
Optimizer for finding the fishing mortality that leads to TAC
}
