% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.smsR.R, R/plotsmsR.R
\name{plot.sms}
\alias{plot.sms}
\title{Plot standard stock assessment figures}
\usage{
\method{plot}{sms}(sas, type = "default", Blim = sas$dat$betaSR, Bpa = NULL, printFig = TRUE)

\method{plot}{sms}(sas, type = "default", Blim = sas$dat$betaSR, Bpa = NULL, printFig = TRUE)
}
\arguments{
\item{sas}{estimated model.}

\item{type}{default, stack, or wrap}

\item{Blim}{optional value to graph if type is stack or wrap}

\item{Bpa}{optional value to add Bpa (Bescape)}

\item{printFig}{}

\item{df.tmb}{list of tmb parameters}
}
\value{
Returns a figure with SSB, Fbar, recruitment and catch

Returns a figure with SSB, Fbar, recruitment and catch
}
\description{
Plot standard stock assessment figures

Plot standard stock assessment figures
}
\examples{
smsPlots(df.tmb, sas)
smsPlots(df.tmb, sas, type="stack", Blim=110000)
smsPlots(df.tmb, sas, type="wrap", Blim=110000)
smsPlots(df.tmb, sas)
smsPlots(df.tmb, sas, type="stack", Blim=110000)
smsPlots(df.tmb, sas, type="wrap", Blim=110000)
}
