% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_OM.R
\name{plot.sms_om}
\alias{plot.sms_om}
\title{Plot standard stock assessment figures}
\usage{
\method{plot}{sms_om}(OM, printFig = 0)
}
\arguments{
\item{OM}{smsR operating model}

\item{printFig}{TRUE FALSE return figure as list}

\item{parms}{list of parameters}
}
\value{
Returns a figure with SSB, Fbar, recruitment and catch
}
\description{
Plot standard stock assessment figures
}
\examples{
smsPlots(df.tmb, sas)
smsPlots(df.tmb, sas, type = "stack", Blim = 110000)
smsPlots(df.tmb, sas, type = "wrap", Blim = 110000)
}
