% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBubbles.R
\name{plotBubbles}
\alias{plotBubbles}
\title{Make a combined bubble plot of survey and catch fits}
\usage{
plotBubbles(sas, CVscale = TRUE)
}
\arguments{
\item{sas}{fitted smsR model}

\item{CVscale}{TRUE/FALSE whether to scale residuals with the estimated CV}
}
\value{
returns a bubble plot
}
\description{
Make a combined bubble plot of survey and catch fits
}
\examples{

plotBubbles(sas)
}
