% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiagnostics.R
\name{plotDiagnostics}
\alias{plotDiagnostics}
\title{Plot all diagnostics of fitted stock assesment model}
\usage{
plotDiagnostics(df.tmb, sas, mr = NULL)
}
\arguments{
\item{sas}{}
}
\description{
Plot all diagnostics of fitted stock assesment model
}
