% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sam_data.R
\name{read.sam.data}
\alias{read.sam.data}
\title{read data from a SAM stock assesment for comparison}
\usage{
read.sam.data(wd)
}
\arguments{
\item{wd}{working directory of sam data}
}
\value{
returns a list of life history, catch, and survey data from SAM
}
\description{
read data from a SAM stock assesment for comparison
}
\examples{
wd <- 'C:/testwd'
read.sam.data('C:/testwd') # Get the list

}
