% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_max_age.R
\name{reduce_max_age}
\alias{reduce_max_age}
\title{Make the plus group smaller than in the data}
\usage{
reduce_max_age(dat, maxage)
}
\arguments{
\item{dat}{data frame containing stock assesssment input}

\item{maxage}{new plus group age}
}
\value{
returns the same data frame
}
\description{
Make the plus group smaller than in the data
}
\examples{
reduce_max_age(dat, maxage = 7)
}
