% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeSurvey.R
\name{removeSurvey}
\alias{removeSurvey}
\title{Remove survey}
\usage{
removeSurvey(sas)
}
\arguments{
\item{sas}{fitted smsR model}
}
\value{
returns a figure of fitted SSB with and without the added surveys
}
\description{
Remove a survey from a fitted `smsR` model, and rerun the assessment. This function will output a figure that compares the SSB with all surveys enabled, and removing them one by one.
}
\examples{
p1 <-  removeSurvey(sas) #
print(p1) # Plot the figure
}
