% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_agebased_sms_OP.R
\name{run.agebased.sms.op}
\alias{run.agebased.sms.op}
\title{Run a full age based model as operating model for SMS}
\usage{
run.agebased.sms.op(df)
}
\arguments{
\item{df}{list of parameters and life history values}
}
\value{
returns a list of derived and calculated values
}
\description{
Run a full age based model as operating model for SMS
}
\examples{
run.agebased.true.catch(df) # runs the model

}
