% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAssessment.R
\name{runAssessment}
\alias{runAssessment}
\title{Fit an smsR model}
\usage{
runAssessment(
  df.tmb,
  parms,
  lwr = list(NA),
  upr = list(NA),
  mps = NULL,
  silent = TRUE,
  debug = FALSE,
  dll = "smsR"
)
}
\arguments{
\item{df.tmb}{A list generated by the \link{get_tmb_parameters} containing model input data and configuration}

\item{parms}{A list or object generated by the \link{getParms} function, containing
model-specific parameters necessary for calculations.}

\item{lwr}{optional list of user evaluated lower bounds}

\item{upr}{optional list of user evaluated lower bounds}

\item{mps}{optional list of parameters to be mapped}

\item{silent}{If `TRUE`, suppresses output messages; if `FALSE`, provides verbose output during execution. Default is `TRUE`.}

\item{debug}{Use debugger to report output from initial parameters}

\item{dll}{Optional. A string specifying the dynamic linked library (DLL) file to use for custom model configurations. Useful for specialized cases.}
}
\value{
An smsR object containing:
\describe{
  \item{\code{opt}}{A list of estimated parameters from \link{nlminb} function.}
  \item{\code{reps}}{Output from the \link{sdreport} TMB function that contains standard deviations and hessian of model fit}
  \item{\code{obj}}{TMB object from \link{MakeADFun}. Contains the output using the initial parameter set}
  \item{\code{dat}}{Input data file \link{df.tmb} for export}
}
}
\description{
run an smsR stock assessment with the data and the estimated parameters generated from \link{get_TMB_parameters} and \link{getParms} respectively.
}
\details{
This function is a flexible interface for fitting stock assessment models using Template Model Builder (TMB). Users can customize parameter bounds, map specific parameters, and specify custom DLL files for advanced configurations. Diagnostics help ensure that the model converged properly and provide insights into the quality of the fit.
The output from \link{runAssessment} can be used to retrieve estimated parameters and derived variables using other `smsR` functions.
}
\examples{
\dontrun{
# Example usage:
df.tmb <- get_tmb_parameters(sandeel1r)
parms <- getParms(df.tmb)
results <- fit_stock_assessment(df.tmb, parms, silent = FALSE)
plot(results)

}

}
\seealso{
\code{\link{getSSB}}, \code{\link{getR}}, \code{\link{getCatch}}, \code{\link{getF}}
}
