% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-OM.R
\name{simple_projection}
\alias{simple_projection}
\title{Do a simple 1 year projection of the stock}
\usage{
simple_projection(df.tmb, N_current, F0, M, weca, mat, ag_years)
}
\arguments{
\item{df.tmb}{list of sms parameters}

\item{N_current}{Numbers at age of the start of the forecast year}

\item{F0}{Fishing mortality}

\item{M}{Natural mortality}

\item{weca}{weight at age in catch}

\item{mat}{Maturity ogive}

\item{west}{weight at age in stock}
}
\value{
returns a list of derived values from the projection
}
\description{
Do a simple 1 year projection of the stock
}
\examples{
# Not run
simple_projection(df.tmb, N_current, F0, M, weca, mat)

}
