% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smsPlots.R
\name{smsPlots}
\alias{smsPlots}
\title{Plot standard stock assessment figures}
\usage{
smsPlots(df.tmb, sas, type = "default", Blim = NULL, printFig = TRUE)
}
\arguments{
\item{df.tmb}{list of tmb parameters}

\item{sas}{estimated model.}

\item{type}{default, stack, or wrap}

\item{Blim}{optional value to graph if type is stack or wrap}
}
\description{
Plot standard stock assessment figures
}
\examples{
smsPlots(df.tmb, sas)
smsPlots(df.tmb, sas, type="stack", Blim=110000)
smsPlots(df.tmb, sas, type="wrap", Blim=110000)
}
