% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{AIC.sms}
\alias{AIC.sms}
\title{Calculate AIC from smsR stock assessment object}
\usage{
\method{AIC}{sms}(sas, p = 2, n = Inf)
}
\arguments{
\item{sas}{Stock assesment object from smsR}

\item{p}{penalty on number of parameters, default = 2}

\item{n}{AICc sample size, default = Inf}
}
\description{
Calculate AIC from smsR stock assessment object
}
