% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{calcBescape}
\alias{calcBescape}
\title{Calculate the fishing mortality required to get the assigned TAC}
\usage{
calcBescape(Btarget, df.tmb, N_current, Fsel, Fcap = 2)
}
\arguments{
\item{Btarget}{Target spawning biomass to reach}

\item{df.tmb}{SMS input list}

\item{N_current}{numbers at age at the beginning of the year}

\item{Fsel}{Current selectivity}

\item{Fcap}{maximum fishing mortality}
}
\description{
Calculate the fishing mortality required to get the assigned TAC
}
