% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{calcFTAC}
\alias{calcFTAC}
\title{Calculate the fishing mortality required to get the assigned TAC}
\usage{
calcFTAC(TAC, df.tmb, N_current, Fsel, Fcap = 2)
}
\arguments{
\item{TAC}{TAC were are trying to achieve}

\item{df.tmb}{SMS input list}

\item{N_current}{numbers at age at the beginning of the year}

\item{Fsel}{Selectivity}

\item{Fcap}{maximum fishing mortality allowed}
}
\value{
Returns the fishing mortality required to reach the input TAC
}
\description{
Calculate the fishing mortality required to get the assigned TAC
}
\examples{

calcFTAC(10000, df.tmb,

}
