% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{calcTAC}
\alias{calcTAC}
\title{Forecast an sms model}
\usage{
calcTAC(
  df.tmb,
  sas,
  recruitment = "hockey",
  HCR = "Fmsy",
<<<<<<< HEAD
  avg_R = NULL,
  Bpa = NULL
=======
  avg_years = rep(3, 4)
>>>>>>> 15619a876e060ce37d66251d5c55c875e90f35a0
)
}
\arguments{
\item{df.tmb}{sms dataframe}

\item{sas}{fitted sms model}

\item{recruitment}{hockey, long_mean or short_mean}

\item{HCR}{Fmsy or Bescape}

<<<<<<< HEAD
\item{avg_R}{years to average R}
=======
\item{avg_years}{years to average weca, M, west, and mat}
>>>>>>> 15619a876e060ce37d66251d5c55c875e90f35a0

\item{method}{}
}
\description{
Forecast an sms model
}
