% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_to_matrix.R
\name{df_to_matrix}
\alias{df_to_matrix}
\title{Turn data frames into multidimensional matrices for TMB}
\usage{
df_to_matrix(df.in, season = 1:4)
}
\arguments{
\item{df.in}{data frame with info}

\item{season}{number of seasons to standardize over}
}
\value{
returns a matrix
}
\description{
Turn data frames into multidimensional matrices for TMB
}
\examples{
df_to_matrix(sandeel_1r$survey, season = 1:2)
}
