% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast.sms}
\alias{forecast.sms}
\title{Simple one year forecast}
\usage{
forecast.sms(df.tmb, N_current, F0)
}
\arguments{
\item{df.tmb}{smsR input data list}

\item{N_current}{Latest estimate of numbers at age. Can be extracted from 'sas' object}

\item{F0}{Fishing mortality and selectivity to project}
}
\value{
returns a list with forecasted SSB and catch
}
\description{
Simple one year forecast
}
\examples{

N_current <-  N_current <- matrix(0 , nrow = df.tmb$nage, ncol = df.tmb$nseason)
N_current[2:df.tmb$nage,1] <- N_temp[2:df.tmb$nage]
N_current[1, df.tmb$recseason] <- N_temp[1]

fcast <- forecast.sms(df.tmb, N_current, rep(0, df.tmb$nage))

}
