% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_OM_parameters.R
\name{get_OM_parameters}
\alias{get_OM_parameters}
\title{Ready parameters for an operating model using a fitted sms model}
\usage{
get_OM_parameters(
  df.tmb,
  sas = NULL,
  surveySD = 0.4,
  recruitment = NULL,
  nspace = 1,
  moveinit = 1,
  movemaxinit = 0.3,
  moveslope = 0.7,
  movefifty = 1
)
}
\arguments{
\item{recruitment}{}
}
\description{
Ready parameters for an operating model using a fitted sms model
}
