% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_OM_parameters.R
\name{get_OM_parameters}
\alias{get_OM_parameters}
\title{Prepare parameters for an operating model using a fitted sms model}
\usage{
get_OM_parameters(
  df.tmb,
  sas = NULL,
  surveySD = 0.4,
  recruitment = NULL,
  nspace = 1,
  moveinit = 1,
  movemax = 0.3,
  rec.space = 1,
  moveslope = 0.7,
  movefifty = 1
)
}
\arguments{
\item{df.tmb}{list of smsR input data}

\item{sas}{fitted smsR stock assessment}

\item{surveySD}{CV on survey}

\item{recruitment}{Type of recruitment function shape}

\item{nspace}{number of spatial cells}

\item{moveinit}{Initial distribution in spatial cells (must add to 1)}

\item{movemax}{maximum movement rate per cell}

\item{rec.space}{relative recruitment in cells}

\item{moveslope}{slope of the movement function shape}

\item{movefifty}{age at 50 percent movement rate out of a cell}
}
\value{
a list of parameters for the OM
}
\description{
Prepare parameters for an operating model using a fitted sms model
}
