% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tmb_parameters.R
\name{get_TMB_parameters}
\alias{get_TMB_parameters}
\title{Organize TMB parameters in a list.}
\usage{
get_TMB_parameters(
  mtrx = NA,
  Surveyobs = NA,
  Catchobs = NA,
  years,
  nseason = 4,
  nsurvey = 2,
  ages = 0:20,
  recseason = 1,
  Fminage = 0,
  Fmaxage = max(ages),
  Qminage = rep(0, length(ages)),
  Qmaxage = rep(max(ages), length(ages)),
  Qlastage = Qmaxage,
  isFseason = rep(1, nseason),
  CminageSeason = rep(0, nseason),
  CmaxageSeason = rep(max(ages), nseason),
  endFseason = 4,
  nocatch = matrix(rep(1, nseason), nrow = length(years), ncol = nseason),
  useEffort = 0,
  effort = matrix(1, nrow = length(years), ncol = nseason),
  blocks = FALSE,
  surveyStart = rep(0, nsurvey),
  surveyEnd = rep(1, nsurvey),
  surveySeason = rep(1, nsurvey),
  powers = list(NA),
  surveyCV = matrix(c(0, max(ages)), nrow = 2, ncol = nsurvey),
  catchCV = matrix(c(0, max(ages)), nrow = 2, ncol = nseason),
  recmodel = 2,
  estCV = c(0, 0, 0),
  CVmin = c(0.2, 0.2, 0.2),
  betaSR = NA,
  nllfactor = rep(1, 3)
)
}
\arguments{
\item{mtrx}{matrix containing M, Mat, weca, west}

\item{Surveyobs}{matrix of survey observations}

\item{Catchobs}{matrix of catch observations}

\item{years}{vector of years}

\item{nseason}{number of seasons}

\item{nsurvey}{number of surveys}

\item{ages}{vector of ages}

\item{recseason}{season where recruitment occurs}

\item{Fminage}{First age to start fishing}

\item{Fmaxage}{Last age with unique fishing mortality}

\item{Qminage}{vector length of nsurvey with minimum ages}

\item{Qmaxage}{vector length of nsurvey with max age with unique selectivity}

\item{Qlastage}{Last age observed in the survey}

\item{isFseason}{Season to calculate F in}

\item{CminageSeason}{Minimum age with fishing mortality per season}

\item{endFseason}{Last season last year with F}

\item{nocatch}{Matrix size year x season with 1 or 0 showing catch}

\item{useEffort}{1 or 0, use nominal CPUE to tune F}

\item{effort}{matrix size year x season that contains nominal CPUE}

\item{blocks}{Blocks with unique selectivity for fishing}

\item{surveyStart}{number between 0 and 1 determine the start of the survey within season}

\item{surveyEnd}{number between 0 and 1 determine the end of the survey within season}

\item{surveySeason}{vector determining which seasons surveys occur}

\item{powers}{which ages are included in the power law calc}

\item{surveyCV}{list with ages with unique survey CVs. Length = nsurvey}

\item{catchCV}{list with ages with unique catch CVs. Length = nseason}

\item{recmodel}{chose recruitment model. 2 = hockeystick}

\item{estCV}{vector length 3 of which CVs to determine. 1-survey,2-catch, 3-stock recruitment}

\item{CVmin}{Currently not functional}

\item{betaSR}{breakpoint of SSB}

\item{nllfactor}{vector length 3 of weights of log likelihood. 1-survey, 2-catch,3-stock recruitment}
}
\value{
list of stuff
}
\description{
Organize TMB parameters in a list.
}
