% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tmb_parameters.R
\name{get_TMB_parameters}
\alias{get_TMB_parameters}
\title{Prepare Input Data for an smsR Model}
\usage{
get_TMB_parameters(
  mtrx = NULL,
  Surveyobs = NULL,
  Catchobs = NULL,
  propM = NULL,
  propF = NULL,
  years,
  CatchProportions = NULL,
  startYear = min(years),
  endYear = max(years),
  nseason = 4,
  nsamples = matrix(0, ncol = nseason, nrow = length(years)),
  nsurvey = dim(Surveyobs)[3],
  ages = 0:20,
  Fbarage = c(1, max(ages)),
  recseason = 1,
  Fminage = 0,
  Fmaxage = max(ages),
  Qminage = rep(0, nsurvey),
  Qmaxage = rep(max(ages), nsurvey),
  Qlastage = NULL,
  isFseason = NULL,
  cscalar = matrix(1, nseason, nage),
  CminageSeason = rep(0, nseason),
  CmaxageSeason = rep(max(ages), nseason),
  endFseason = nseason,
  nocatch = matrix(rep(1, nseason), nrow = length(years), ncol = nseason),
  useEffort = 0,
  estimateCreep = 0,
  effort = matrix(1, nrow = length(years), ncol = nseason),
  blocks = FALSE,
  surveyStart = rep(0, nsurvey),
  surveyEnd = rep(1, nsurvey),
  surveySeason = rep(1, nsurvey),
  leavesurveyout = rep(1, nsurvey),
  minSDsurvey = 0.1,
  minSDcatch = 0.1,
  maxSDcatch = sqrt(2),
  peneps = 0.001,
  penepsC = 0.001,
  penepsCmax = 0.001,
  Mprior = -1,
  SDMprior = 0.2,
  powers = rep(list(NA), nsurvey),
  Pred_in = matrix(0),
  M_min = 0,
  M_max = max(ages),
  scv = array(0, dim = c(length(ages), length(years), nsurvey)),
  surveySD = replicate(nsurvey, 0, simplify = FALSE),
  catchSD = replicate(nseason, 0, simplify = FALSE),
  MCV = matrix(c(0, max(ages)), nrow = 2, ncol = 1),
  csd_break = min(years),
  csd = 0,
  tuneCatch = 0,
  tuneStart = NULL,
  tuneEnd = NULL,
  prior_SDM = 0.4,
  estSD = c(0, 0, 0),
  SDmin = c(0.2, 0.2, 0.2),
  betaSR = NULL,
  nllfactor = rep(1, 3),
  randomF = 0,
  randomM = 0,
  randomR = 0,
  recmodel = 1,
  nenv = 0,
  warn = FALSE
)
}
\arguments{
\item{mtrx}{Matrix containing life history parameters: `weca`, `west`, `mat`, and `M`.}

\item{Surveyobs}{Matrix of survey observations (numbers at age).}

\item{Catchobs}{Matrix of catch observations (numbers at age).}

\item{propM}{Proportion of natural mortality (`M`) occurring before spawning.}

\item{propF}{Proportion of fishing mortality (`F`) occurring before spawning.}

\item{years}{Vector of years with available data.}

\item{startYear}{Optional. Start year of the assessment.}

\item{endYear}{Optional. End year of the assessment.}

\item{nseason}{Number of seasons in the model.}

\item{nsurvey}{Number of surveys included in the model.}

\item{ages}{Vector specifying the age span of the data.}

\item{Fbarage}{Vector of ages used to calculate the mean fishing mortality (`Fbar`).}

\item{recseason}{Season in which recruitment occurs.}

\item{Fminage}{Minimum age used to calculate fishing mortality (`F`).}

\item{Fmaxage}{Maximum age used to calculate fishing mortality (`F`).}

\item{Qminage}{Minimum age included in the survey data.}

\item{Qmaxage}{Maximum age included in the survey data.}

\item{Qlastage}{Last age with unique survey selectivity.}

\item{isFseason}{Vector specifying seasons in which fishing mortality (`F`) is calculated.}

\item{CminageSeason}{Minimum age for using catch data.}

\item{CmaxageSeason}{Maximum age for using catch data.}

\item{endFseason}{Last season in which fishing occurs.}

\item{nocatch}{Matrix (1/0) defining whether fishing mortality (`F`) should be calculated.}

\item{useEffort}{Logical. If `TRUE`, nominal effort is used to calculate `F`.}

\item{estimateCreep}{Logical. If `TRUE`, estimates technological creep as a parameter (requires `useEffort = TRUE`).}

\item{effort}{Matrix of nominal fishing effort per season.}

\item{blocks}{Vector of unique selectivity blocks.}

\item{surveyStart}{Fraction into the season when the survey starts.}

\item{surveyEnd}{Fraction into the season when the survey ends.}

\item{surveySeason}{Season in which the survey occurs.}

\item{leavesurveyout}{Vector (1/0) indicating which surveys to exclude.}

\item{minSDsurvey}{Minimum coefficient of variation (CV) for survey data.}

\item{minSDcatch}{Minimum CV for catch data.}

\item{maxSDcatch}{Maximum CV for catch data.}

\item{peneps}{Parameter regulating the minimum CV for surveys.}

\item{penepsC}{Parameter regulating the minimum CV for catch data.}

\item{penepsCmax}{Parameter regulating the maximum CV for catch data.}

\item{Mprior}{Prior standard deviation (SD) for `M` deviations from the first year of the time series. Negative value indicates this it is not used.}

\item{SDMprior}{Prior SD for `M` CV estimation.}

\item{powers}{Logical. If `TRUE`, applies a power-law function for density-dependent survey observations.}

\item{Pred_in}{Matrix of predator inputs for MICE modeling.}

\item{M_min}{Minimum age included in the time varying `M` estimation.}

\item{M_max}{Maximum age included in the time varying `M` estimation.}

\item{scv}{Matrix specifying time-varying survey CV.}

\item{surveySD}{Grouping variable for survey SD.}

\item{catchSD}{Grouping variable for catch SD.}

\item{MCV}{Age distribution of the CV for time-varying `M`.}

\item{prior_SDM}{Prior for `M` estimation}

\item{estSD}{Vector indicating which deviation parameters to estimate.}

\item{SDmin}{Minimum CV values for estimation.}

\item{betaSR}{Hockey-stick model breakpoint parameter.}

\item{nllfactor}{Weighting factor for the negative log-likelihood.}

\item{randomF}{Logical. If `TRUE`, fishing mortality (`F`) is treated as a random effect.}

\item{randomM}{Logical. If `TRUE`, natural mortality (`M`) is estimated as a time-varying random effect.}

\item{randomR}{Logical. If `TRUE`, recruitment deviations are estimated as a random effect.}

\item{recmodel}{Recruitment model specification. Options:
- `1`: Hockey-stick model with input breakpoint (`betaSR`).
- `2`: Recruitment estimated as deviations from an estimated mean value with an AR1 process in the likelihood
- `3`: Beverton Holt model with steepness.
- `4`: Beverton Holt model with steepness and environmentally mediated deviations - requires env_matrix}

\item{nenv}{Number of environmental covariates affecting recruitment.}
}
\value{
A list containing all required input data for the smsR model.
}
\description{
This function prepares the necessary input data for an `smsR` model.
 Required input are `mtrx`, `Surveyobs`,`Catchobs`, and `year`.
 Note that the dimensions of the input data needs to be consistent with year, season and ages.
}
\details{
The required inputs include:
  - `mtrx`: A matrix containing life history parameters (i.e,, `weca`, `west`, `mat`, `M`).
  - `Catchobs`: A matrix of observed catch numbers at age.
  - `Surveyobs`: A matrix of observed survey numbers at age.

Additionally, the user must specify the `years` corresponding to the input data.
}
\examples{
get_TMB_parameters(
  mtrx = sandeel_1r$lhs,
  Surveyobs = sandeel_1r$survey,
  Catchobs = sandeel_1r$Catch,
  years = 1983:2022
) # Not run
}
