% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tmb_parameters.R
\name{get_TMB_parameters}
\alias{get_TMB_parameters}
\title{Prepare data input for an smsR model}
\usage{
get_TMB_parameters(
  mtrx = NULL,
  Surveyobs = NULL,
  Catchobs = NULL,
  propM = NULL,
  propF = NULL,
  years,
  startYear = min(years),
  endYear = max(years),
  nseason = 4,
  nsurvey = 2,
  ages = 0:20,
  Fbarage = c(1, max(ages)),
  recseason = 1,
  Fminage = 0,
  Fmaxage = max(ages),
  Qminage = rep(0, length(ages)),
  Qmaxage = rep(max(ages), length(ages)),
  Qlastage = Qmaxage,
  isFseason = rep(1, nseason),
  CminageSeason = rep(0, nseason),
  CmaxageSeason = rep(max(ages), nseason),
  endFseason = nseason,
  nocatch = matrix(rep(1, nseason), nrow = length(years), ncol = nseason),
  useEffort = 0,
  estimateCreep = 0,
  effort = matrix(1, nrow = length(years), ncol = nseason),
  blocks = FALSE,
  surveyStart = rep(0, nsurvey),
  surveyEnd = rep(1, nsurvey),
  surveySeason = rep(1, nsurvey),
  leavesurveyout = rep(1, nsurvey),
  minSDsurvey = 0.3,
  minSDcatch = 0.2,
  maxSDcatch = sqrt(2),
  peneps = 0.001,
  penepsC = 0.001,
  penepsCmax = 0.001,
  Mprior = 0.5,
  powers = list(NA),
  Pred_in = NA,
  M_min = 0,
  M_max = max(ages),
  scv = array(0, dim = c(length(ages), length(years), nsurvey)),
  surveyCV = matrix(c(0, max(ages)), nrow = 2, ncol = nsurvey),
  catchCV = matrix(c(0, max(ages)), nrow = 2, ncol = nseason),
  MCV = matrix(c(0, max(ages)), nrow = 2, ncol = 1),
  estCV = c(0, 0, 0),
  CVmin = c(0.2, 0.2, 0.2),
  betaSR = 0,
  nllfactor = rep(1, 3),
  randomF = 0,
  randomM = 0,
  randomR = 0,
  recmodel = 1,
  nenv = 0
)
}
\arguments{
\item{mtrx}{matrix containing weca, west, mat, and M}

\item{Surveyobs}{Survey observations}

\item{Catchobs}{Catch observations}

\item{propM}{proportion M before spawning}

\item{propF}{proportion F before spawning}

\item{years}{years with available data}

\item{startYear}{start year of assessment (optional)}

\item{endYear}{end year of assessment (optional)}

\item{nseason}{number of seasons}

\item{nsurvey}{number of surveys}

\item{ages}{age span}

\item{Fbarage}{Ages to average F to Fbar}

\item{recseason}{recruitment season}

\item{Fminage}{minimum age to calculate F}

\item{Fmaxage}{Maximum age to calculate F}

\item{Qminage}{Minimum age included in survey}

\item{Qmaxage}{Max age included in survey}

\item{Qlastage}{Last age with unique survey selectivity}

\item{isFseason}{seasons to calculate F}

\item{CminageSeason}{minimum age to use catchdata}

\item{CmaxageSeason}{Maximum age to use catchdata}

\item{endFseason}{Last season with fishing}

\item{nocatch}{input matrix of 1 or 0 defining if F should be calculated}

\item{useEffort}{Use nominal effort to calculate F}

\item{estimateCreep}{Estimate technological creep as a parameter. Needs useEffort = 1}

\item{effort}{matrix of nominal effort per season}

\item{blocks}{unique selectivity blocks}

\item{surveyStart}{fraction into the season where the survey starts}

\item{surveyEnd}{fraction into the season where the survey ends}

\item{surveySeason}{season where survey occurs}

\item{leavesurveyout}{vector of 1 and 0 to exclude surveys}

\item{minSDsurvey}{minimum CV of survey}

\item{minSDcatch}{minimum CV of catch}

\item{maxSDcatch}{Maximum CV of catch}

\item{peneps}{parameter for regulating minimum CV for survey}

\item{penepsC}{parameter for regulating minimum CV of catch}

\item{penepsCmax}{parameter for regulating maximum CV of catch}

\item{Mprior}{Prior SD on M deviations from the first year of the time series}

\item{powers}{apply powerlaw for density dependent survey observations}

\item{M_min}{minimum age for time varying M}

\item{M_max}{maximum age for time varying M}

\item{scv}{add time varying survey CV (input matrix)}

\item{surveyCV}{survey cv grouping}

\item{catchCV}{catch cv grouping}

\item{MCV}{Age distribution of time varying M CV}

\item{estCV}{which CVs to estimate as parameters}

\item{CVmin}{minimum CVs}

\item{betaSR}{hockey stick break point}

\item{nllfactor}{negative log likelihood weighting}

\item{randomF}{try random effect fishing mortality}

\item{randomM}{Random effect estimating time varying M}

\item{randomR}{Estimate R deviations as a random efffect}

\item{recmodel}{recruitment model}

\item{nenv}{Number of environemntal covariates on  R}

\item{nalphaM}{Number of external predators}
}
\value{
returns a
}
\description{
Prepare data input for an smsR model
}
\examples{
get_TMB_paramters(
  mtrx = sandeel_1r$lhs,
  Surveyobs = sandeel_1r$survey,
  Catchobs = sandeel_1r$Catch,
  years = 1983:2022
) # Not run

}
