% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getBiomass}
\alias{getBiomass}
\title{Get a data frame of the biomass of each age group}
\usage{
getBiomass(df.tmb, sas)
}
\arguments{
\item{df.tmb}{list of input parameters}

\item{sas}{fitted smsR model}
}
\value{
data frame containing the biomass of each age group.
low and high are the 95\% confidence intervals.
SE is standard error of log Biomass
}
\description{
Get a data frame of the biomass of each age group
}
\seealso{
[getR()],[getYield()],[getF()],[getCatch()], [getSSB()], [getN()], [getFbar()], [getCatchN()], [getCatchCV()],
[getSurvey()]
}
