% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getCatch}
\alias{getCatch}
\title{Retrieve the total catch from a fitted smsR object}
\usage{
getCatch(df.tmb, sas)
}
\arguments{
\item{df.tmb}{input parameters}

\item{sas}{fitted smsR object}
}
\value{
returns a data frame
}
\description{
Retrieve the total catch from a fitted smsR object
}
\examples{
catch <- getCatch(df.tmb, sas)
print(catch)
}
