% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getCatchability}
\alias{getCatchability}
\title{Get the survey catchability}
\usage{
getCatchability(df.tmb, sas)
}
\arguments{
\item{df.tmb}{sms input parameters}

\item{sas}{fitted model object}
}
\description{
Get the survey catchability
}
