% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getCatchCV}
\alias{getCatchCV}
\title{Retrieve estimated catch CV}
\usage{
getCatchCV(df.tmb, sas)
}
\arguments{
\item{df.tmb}{Input list for smsR data}

\item{sas}{stock assessment output from smsR}
}
\value{
data frame of estimated parameters and their SE
}
\description{
Retrieve estimated catch CV
}
