% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataSMS.R
\name{getDataSMS}
\alias{getDataSMS}
\title{Read data from an SMS model}
\usage{
getDataSMS(
  wd,
  maxage = NA,
  survey.age = list(),
  survey.years = list(),
  survey.names = NA,
  survey.quarter = NA,
  effort = FALSE,
  years,
  seasons,
  scv.tv = 0
)
}
\arguments{
\item{wd}{working directory}

\item{maxage}{maxage}

\item{survey.age}{list showing ages in surveys}

\item{survey.years}{years ssurveys occur}

\item{survey.names}{names of survey}

\item{survey.quarter}{which quarters do surveys happen}

\item{effort}{set to true to read in effort data}

\item{years}{years to calculate}

\item{seasons}{seasons to calc}

\item{scv.tv}{Time varying CV from survey}
}
\value{
returns a list of input data for smsR from an old sms format
}
\description{
Read data from an SMS model
}
