% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getEstimatedParms}
\alias{getEstimatedParms}
\title{Estimated parameters}
\usage{
getEstimatedParms(sas)
}
\arguments{
\item{sas}{stock assessment output from smsR}
}
\value{
data frame of estimated parameters and their SE
}
\description{
Retrieve estimated parameters from a fitted `smsR` model with their associated uncertainty.
}
\examples{

print(getEstimatedParms(sas))

}
