% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getFbar}
\alias{getFbar}
\title{Estimated Fbar}
\usage{
getFbar(df.tmb, sas)
}
\arguments{
\item{df.tmb}{list of input parameters}

\item{sas}{smsR fitted model}
}
\value{
data frame containing the average fishing mortality each year.
}
\description{
Get the estimated `Fbar` (age averaged fishing mortality) from an fitted `smsR` model.
}
\details{
low and high are the 95\% confidence intervals.
SE is standard error of log Fbar
Fbar depends on the age ranged specified in the  \code{\link{get_TMB_parameters}} call for `df.tmb`
}
\examples{

Fbar <- getFbar(df.tmb, sas)

}
