% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getFbar}
\alias{getFbar}
\title{Get Fbar from a fitted smsR model}
\usage{
getFbar(df.tmb, sas)
}
\arguments{
\item{df.tmb}{list of input parameters}

\item{sas}{smsR fitted model}
}
\value{
data frame containing the average fishing mortality each year.
low and high are the 95\% confidence intervals.
SE is standard error of log Fbar
}
\description{
Get Fbar from a fitted smsR model
}
\examples{

Fbar <- getFbar(df.tmb, sas)

}
