% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFmsy.R
\name{getFmsy}
\alias{getFmsy}
\title{Calculate Fmsy from a sms model}
\usage{
getFmsy(df.tmb, sas, nyears = 50, recruitment = "hockey", plotMSY = FALSE)
}
\arguments{
\item{df.tmb}{sms data frame}

\item{sas}{Fitted sms model}

\item{nyears}{number of years to simulate}

\item{recruitment}{recruitment type to calculate Fmsy. Currently only hockey works.}

\item{plotMSY}{TRUE OR FALSE to plot the result}
}
\description{
Calculate Fmsy from a sms model
}
