% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFmsy.R
\name{getFmsy}
\alias{getFmsy}
\title{Calculate Fmsy from a sms model}
\usage{
getFmsy(
  sas,
  nyears = 50,
  recruitment = "hockey",
  plotMSY = FALSE,
  stochastic = FALSE
)
}
\arguments{
\item{sas}{Fitted sms model}

\item{nyears}{number of years to simulate}

\item{recruitment}{recruitment type to calculate Fmsy. Currently only hockey works.}

\item{plotMSY}{TRUE OR FALSE to plot the result}

\item{stochastic}{stochastic fmsy?}

\item{df.tmb}{sms data frame}
}
\value{
returns the fishing mortality that leads to Fmsy
}
\description{
Calculate Fmsy from a sms model
}
