% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{getForecastTable}
\alias{getForecastTable}
\title{Produce an ICES forecast table}
\usage{
getForecastTable(
  df.tmb,
  sas,
  TACold,
  HCR = "Bescape",
  avg_R = NULL,
  recruitment = "mean",
  Btarget = NULL,
  TACtarget = NULL,
  Fcap = 2
)
}
\arguments{
\item{df.tmb}{sms model input data}

\item{sas}{fitted sms model}

\item{TACold}{last years tac}

\item{HCR}{Harvest control rule to use - options Bescape or Fmsy}

\item{avg_R}{years to average recruitment}

\item{recruitment}{projected recruitment option ('mean)}

\item{Btarget}{SSB at Btarget}

\item{Fcap}{Maximum possible F value}
}
\value{
Returns a table with different catch options
}
\description{
Produce an ICES forecast table
}
\examples{
getForecastTable(df.tmb, sas, TACold = 1e4, HCR = 'Bescape')

}
