% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getN}
\alias{getN}
\title{Get numbers at age from a fitted model}
\usage{
getN(df.tmb, sas)
}
\arguments{
\item{df.tmb}{list of input parameters}

\item{sas}{fitted smsR model}
}
\value{
data frame of Numbers at age.
Last year is based on the expected survival from terminal year.
low and high are the 95\% confidence intervals.
SE is standard error of log N
}
\description{
Get numbers at age from a fitted model
}
\examples{
N <- getN(df.tmb, sas)
}
