% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParms.R
\name{getParms}
\alias{getParms}
\title{Title}
\usage{
getParms(df.tmb, parms.true = NULL)
}
\arguments{
\item{df.tmb}{input values for TMB file}

\item{parms.true}{data frame of true parameters}
}
\value{
returns a set of parameters to estimate for an smsR model
}
\description{
Title
}
\examples{

parms <- getParms(df.tmb)

}
