% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getR}
\alias{getR}
\title{Get a data frame of recruitment from a fitted model}
\usage{
getR(df.tmb, sas)
}
\arguments{
\item{df.tmb}{list of input parameters}

\item{sas}{fitted smsR model}
}
\value{
data frame of recruitment.
Last year is based on the Stock recruitment relationship.
low and high are the 95\% confidence intervals.
SE is standard error of log R
}
\description{
Get a data frame of recruitment from a fitted model
}
\examples{
R <- getR(df.tmb, sas)
print(R)
}
