% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getResidCatch}
\alias{getResidCatch}
\title{Get the catch residuals}
\usage{
getResidCatch(df.tmb, sas)
}
\arguments{
\item{df.tmb}{list of input data for smsR object}

\item{sas}{fitted stock assessment model for smsR}
}
\value{
Returns a data frame of catch residuals
}
\description{
Get the catch residuals
}
\examples{

getResidCatch(df.tmb, sas)
print(getResidCatch)

}
