% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getResidSurvey}
\alias{getResidSurvey}
\title{Get survey residuals from fitted model}
\usage{
getResidSurvey(df.tmb, sas)
}
\arguments{
\item{df.tmb}{list of input parameters}

\item{sas}{fitted smsR model}
}
\value{
data frame of fishing mortality at age. Last year is based on the expected survival from terminal year. low and high are the 95\% confidence intervals
}
\description{
Get survey residuals from fitted model
}
\examples{
res <- getResidSurvey(df.tmb, sas)
print(res)

}
