% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getSR}
\alias{getSR}
\title{Get the stock recruitment relationship}
\usage{
getSR(df.tmb, sas)
}
\arguments{
\item{df.tmb}{data frame with input data}

\item{sas}{fitted smsR model}
}
\value{
returns a data frame with a stock recruitment relationship
}
\description{
Get the stock recruitment relationship
}
\examples{

SR <- getSR(df.tmb, sas)

}
