% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getSSB}
\alias{getSSB}
\title{Get a data frame of the spawning stock biomass}
\usage{
getSSB(df.tmb, sas)
}
\arguments{
\item{df.tmb}{list of input parameters}

\item{sas}{fitted smsR model}
}
\value{
data frame of spawning biomass. minSE and maxSE is the 95% confidence intervals, se is the standard error
}
\description{
Get a data frame of the spawning stock biomass
}
