% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getSSB}
\alias{getSSB}
\title{Get a data frame of the spawning stock biomass}
\usage{
getSSB(df.tmb, sas)
}
\arguments{
\item{df.tmb}{list of input parameters}

\item{sas}{fitted smsR model}
}
\value{
data frame of spawning stock biomass and 95\% confidence intervals.
SE is standard error of log SSB
}
\description{
`getSSB()` retrieves the spawning stock biomass per year, and it's associated uncertainty in the 95th confidence region
}
\examples{
SSB <- getSSB(df.tmb, sas)
print(SSB)
}
\seealso{
[getR()] why are the brackets here
}
