% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getSummary}
\alias{getSummary}
\title{Get summary of derived variables from fitted smsR object}
\usage{
getSummary(df.tmb, sas, Fbarage = NULL)
}
\arguments{
\item{Fbarage}{}
}
\description{
Get summary of derived variables from fitted smsR object
}
