% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getSummaryCVs}
\alias{getSummaryCVs}
\title{Get the CVs of the summary information}
\usage{
getSummaryCVs(df.tmb, sas, verbose = TRUE)
}
\arguments{
\item{df.tmb}{list of smsR input data}

\item{sas}{fitted smsR model}

\item{verbose}{print the output}
}
\value{
returns a table of annual CVs for SSB, R and Fbar
}
\description{
Get the CVs of the summary information
}
\examples{

cvs <- getSummaryCVs(df.tmb, sas, verbose = TRUE) # Don't print it

}
