% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{getTAC}
\alias{getTAC}
\title{Forecast an sms model}
\usage{
getTAC(
  df.tmb,
  sas,
  recruitment = "mean",
  HCR = "Bescape",
  avg_R = NULL,
  Btarget = NULL,
  Fcap = 2
)
}
\arguments{
\item{df.tmb}{sms dataframe}

\item{sas}{fitted sms model}

\item{recruitment}{hockey, mean}

\item{HCR}{Fmsy or Bescape}

\item{avg_R}{vector of years to average R (e.g. 2010:2022)}

\item{Btarget}{SSB value for Btarget}

\item{Fcap}{Maximum possible fishing mortality}
}
\value{
returns the TAC based on the applied HCR
}
\description{
Forecast an sms model
}
\examples{
getTAC(df.tmb, sas, HCR = 'Bescape')

}
