% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getTSB}
\alias{getTSB}
\title{Get the total stock biomass}
\usage{
getTSB(df.tmb, sas)
}
\arguments{
\item{df.tmb}{list of input parameters}

\item{sas}{fitted smsR model}
}
\value{
data frame containing the biomass of each age group.
low and high are the 95\% confidence intervals.
SE is standard error of log TSB
}
\description{
Get the total stock biomass
}
\examples{

dat <- getTSB(df.tmb, sas)

}
\seealso{
[getR()],[getSSB()],[getYield()],[getF()],[getCatch()], [getBiomass()], [getN()], [getFbar()], [getCatchN()], [getCatchCV()],
[getSurvey()]
}
