% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-output.R
\name{getYield}
\alias{getYield}
\title{Get the observed yield from the fishery}
\usage{
getYield(df.tmb)
}
\arguments{
\item{df.tmb}{a list of parameters for smsR model}
}
\value{
returns the observed catch at age
}
\description{
Get the observed yield from the fishery
}
\examples{

Yield <- getYield(df.tmb) # Get observed catch at age as a data frame

}
