% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mohns_rho.R, R/mohns_vs2.R
\name{mohns_rho}
\alias{mohns_rho}
\title{Calculate Mohns \eqn{\rho}.}
\usage{
mohns_rho(
  df.tmb,
  parms,
  peels = 5,
  mps = NULL,
  useSSBprojection = TRUE,
  lwr = list(NA),
  upr = list(NA),
  plotfigure = TRUE,
  limits = NULL,
  dll = "smsR"
)

mohns_rho(
  df.tmb,
  parms,
  peels = 5,
  mps = NULL,
  useSSBprojection = TRUE,
  lwr = list(NA),
  upr = list(NA),
  plotfigure = TRUE,
  limits = NULL,
  dll = "smsR"
)
}
\arguments{
\item{df.tmb}{list of tmb parameters}

\item{parms}{parameters that are estimated}

\item{peels}{number of peels}

\item{mps}{parameters that are mapped}

\item{useSSBprojection}{Use the projected SSB to calculate Mohns rho}

\item{plotfigure}{TRUE FALSE of whether to plot figure}

\item{limits}{limits on plot}

\item{dll}{which dll to use. Uses 'sms' as standard}

\item{lower}{lower boundary of parameters}

\item{upper}{upper boundary of parameters}
}
\value{
returns a list of Mohns rho values, input values, and a Mohns rho plot

returns a list of Mohns rho values, input values, and a Mohns rho plot
}
\description{
Calculate the retrospective patterns of a fitted stock assessment model from `smsR`.
The function calculates Mohn's \eqn{\rho} (Mohn 1999) on SSB, Fbar and R, and by default takes 5 peels off.

Calculate the retrospective patterns of a fitted stock assessment model from `smsR`.
The function calculates Mohn's \eqn{\rho} (Mohn 1999) on SSB, Fbar and R, and by default takes 5 peels off.
}
\examples{

MR <- mohns_rho(df.tmb, peels = 5, parms)

print(MR$p1()) # plot the results

MR <- mohns_rho(df.tmb, peels = 5, parms)

print(MR$p1()) # plot the results
}
\references{
Mohn, R. (1999). The retrospective problem in sequential population analysis:
An investigation using cod stock in the Northwest Atlantic. *ICES Journal of Marine Science*, 56(4), 473-488. \url{10.1006/jmsc.1999.0481}

Mohn, R. (1999). The retrospective problem in sequential population analysis:
An investigation using cod stock in the Northwest Atlantic. *ICES Journal of Marine Science*, 56(4), 473-488. \url{10.1006/jmsc.1999.0481}
}
