% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mohns_rho.R
\name{mohns_rho}
\alias{mohns_rho}
\title{Calculate mohns rho.}
\usage{
mohns_rho(
  df.tmb,
  peels = 5,
  parms,
  mps = NULL,
  useSSBprojection = TRUE,
  lwr = list(NA),
  upr = list(NA),
  plotfigure = TRUE,
  limits = NULL,
  dll = "smsR"
)
}
\arguments{
\item{df.tmb}{list of tmb parameters}

\item{peels}{number of peels}

\item{parms}{parameters that are estimated}

\item{mps}{parameters that are mapped}

\item{useSSBprojection}{Use the projected SSB to calculate Mohns rho}

\item{plotfigure}{TRUE FALSE of whether to plot figure}

\item{limits}{limits on plot}

\item{dll}{which dll to use. Uses 'sms' as standard}

\item{lower}{lower boundary of parameters}

\item{upper}{upper boundary of parameters}
}
\value{
returnsa list of Mohns rho values, input values, and a Mohns rho plot
}
\description{
Calculate mohns rho.
}
\examples{

MR <- mohns_rho(df.tmb, peels = 5, parms)

print(MR$p1()) # plot the results
}
