% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_smsR.R
\name{plot.sms}
\alias{plot.sms}
\title{Plot standard stock assessment figures}
\usage{
\method{plot}{sms}(
  sas,
  df.tmb = NULL,
  type = "default",
  Blim = sas$dat$betaSR,
  Bpa = NULL,
  printFig = 0
)
}
\arguments{
\item{sas}{estimated model.}

\item{df.tmb}{list of tmb parameters}

\item{type}{default, stack, or wrap}

\item{Blim}{optional value to graph if type is stack or wrap}

\item{Bpa}{optional value to add Bpa (Bescape)}

\item{printFig}{TRUE FALSE return figure as list}
}
\value{
Returns a figure with SSB, Fbar, recruitment and catch
}
\description{
Plot standard stock assessment figures
}
\examples{
smsPlots(df.tmb, sas)
smsPlots(df.tmb, sas, type = "stack", Blim = 110000)
smsPlots(df.tmb, sas, type = "wrap", Blim = 110000)
}
