% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDiagnostics.R
\name{plotDiagnostics}
\alias{plotDiagnostics}
\title{Plot all diagnostics of fitted stock assesment model}
\usage{
plotDiagnostics(df.tmb, sas, mr = NULL)
}
\arguments{
\item{df.tmb}{List of smsR input data}

\item{sas}{Fitted smsR model}
}
\value{
returns a list of plots
}
\description{
Plot all diagnostics of fitted stock assesment model
}
\examples{
p <- plotDiagnostics(df.tmb, sas)

print(p$SR) # Print the stock recruitment relationship

}
