% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeSurvey.R
\name{removeSurvey}
\alias{removeSurvey}
\title{Plot a comparison of stock assessments when removing surveys}
\usage{
removeSurvey(sas, parms = NULL)
}
\arguments{
\item{sas}{fitted smsR model}

\item{parms}{optional parameters}
}
\value{
returns a figure
}
\description{
Plot a comparison of stock assessments when removing surveys
}
\examples{
p1 <-  removeSurvey(sas) #
print(p1) # Plot the figure
}
