% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAssessment.R
\name{runAssessment}
\alias{runAssessment}
\title{Run the sms model}
\usage{
runAssessment(
  df.tmb,
  parms,
  lwr = list(NA),
  upr = list(NA),
  mps = NULL,
  silent = TRUE
)
}
\arguments{
\item{df.tmb}{list of tmb parameters}

\item{parms}{optional parameter input}

\item{mps}{optional mapping input}

\item{silent}{Supress TMB output}

\item{lower}{lower bounds}

\item{upper}{upper bounds}
}
\value{
a list of derived variables
}
\description{
Run the sms model
}
\examples{
sas <- runAssessment(df.tmb, parms) # Run an assessment
}
