% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAssessment.R
\name{runAssessment}
\alias{runAssessment}
\title{run the assessment with the data and the estimated parameters}
\usage{
runAssessment(
  df.tmb,
  parms,
  lwr = list(NA),
  upr = list(NA),
  mps = NULL,
  silent = TRUE,
  dll = "smsR"
)
}
\arguments{
\item{df.tmb}{A list generated by the `get_tmb_parameters` containing model input data and configuration}

\item{parms}{A list or object generated by the `getParms` function, containing
model-specific parameters necessary for calculations.}

\item{lwr}{optional list of user evaluated lower bounds}

\item{upr}{optional list of user evaluated lower bounds}

\item{mps}{optional list of parameters to be mapped}

\item{silent}{If `TRUE`, suppresses output messages; if `FALSE`, provides verbose output during execution. Default is `TRUE`.}

\item{dll}{Optional. A string specifying the dynamic linked library (DLL) file to use for custom model configurations. Useful for specialized cases.}
}
\value{
A list containing:
\describe{
  \item{\code{fitted_model}}{The fitted stock assessment model object.}
  \item{\code{estimated_params}}{A list of estimated parameters with their values.}
  \item{\code{convergence_status}}{An indicator of whether the optimization converged successfully.}
  \item{\code{diagnostics}}{Diagnostics and additional information about the model fit.}
}
}
\description{
run the assessment with the data and the estimated parameters
}
\details{
This function is a flexible interface for fitting stock assessment models using Template Model Builder (TMB). Users can customize parameter bounds, map specific parameters, and specify custom DLL files for advanced configurations. Diagnostics help ensure that the model converged properly and provide insights into the quality of the fit.
}
\examples{
\dontrun{
# Example usage:
df.tmb <- get_tmb_parameters(sandeel1r)
parms <- getParms(df.tmb)
results <- fit_stock_assessment(df.tmb, parms, silent = FALSE)
plot(results)

}

}
