% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAssessment.R
\name{runAssessment}
\alias{runAssessment}
\title{Run the sms model}
\usage{
runAssessment(df.tmb, lwr = list(NA), upr = list(NA), parms, mps = NULL)
}
\arguments{
\item{df.tmb}{list of tmb parameters}

\item{parms}{optional parameter input}

\item{mps}{optional mapping input}

\item{lower}{lower bounds}

\item{upper}{upper bounds}
}
\value{
a list of derived variables
}
\description{
Run the sms model
}
