% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAssessment.R
\name{runAssessment}
\alias{runAssessment}
\title{run the assessment with the data and the estimated parameters}
\usage{
runAssessment(
  df.tmb,
  parms,
  lwr = list(NA),
  upr = list(NA),
  mps = NULL,
  silent = TRUE,
  debug = FALSE,
  dll = "smsR"
)
}
\arguments{
\item{df.tmb}{A list generated by the \link{get_tmb_parameters} containing model input data and configuration}

\item{parms}{A list or object generated by the \link{getParms} function, containing
model-specific parameters necessary for calculations.}

\item{lwr}{optional list of user evaluated lower bounds}

\item{upr}{optional list of user evaluated lower bounds}

\item{mps}{optional list of parameters to be mapped}

\item{silent}{If `TRUE`, suppresses output messages; if `FALSE`, provides verbose output during execution. Default is `TRUE`.}

\item{debug}{Use debugger to report output from initial parameters}

\item{dll}{Optional. A string specifying the dynamic linked library (DLL) file to use for custom model configurations. Useful for specialized cases.}

\item{save_all}{save all output or keep just SSB, R, Catch, and biomass}
}
\value{
An smsR object containing:
\describe{
  \item{\code{opt}}{A list of estimated parameters from \link{nlminb} function.}
  \item{\code{reps}}{Output from the \link{sdreport} TMB function that contains standard deviations and hessian of model fit}
  \item{\code{obj}}{TMB object from \link{MakeADFun}. Contains the output using the initial parameter set}
  \item{\code{dat}}{Input data file \link{df.tmb}for export}
}
}
\description{
run the assessment with the data and the estimated parameters
}
\details{
This function is a flexible interface for fitting stock assessment models using Template Model Builder (TMB). Users can customize parameter bounds, map specific parameters, and specify custom DLL files for advanced configurations. Diagnostics help ensure that the model converged properly and provide insights into the quality of the fit.
}
\examples{
\dontrun{
# Example usage:
df.tmb <- get_tmb_parameters(sandeel1r)
parms <- getParms(df.tmb)
results <- fit_stock_assessment(df.tmb, parms, silent = FALSE)
plot(results)

}

}
