% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveOutput.R
\name{saveOutput}
\alias{saveOutput}
\title{Save output to evaluate assessment model}
\usage{
saveOutput(
  df.tmb,
  sas,
  MR = NULL,
  save = TRUE,
  Fbarage = c(1, 2),
  name = "summary",
  wd = getwd()
)
}
\arguments{
\item{wd}{}
}
\description{
Save output to evaluate assessment model
}
