% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smsPlots.R
\name{smsPlots}
\alias{smsPlots}
\title{Title}
\usage{
smsPlots(df.tmb, sas)
}
\arguments{
\item{df.tmb}{list of tmb parameters}

\item{sas}{estimated model.}
}
\description{
Title
}
\examples{
smsplots(df.tmb, reps)

}
