% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smsPlots.R
\name{smsPlots}
\alias{smsPlots}
\title{Title}
\usage{
smsPlots(df.tmb, reps, Fbarage = df.tmb$age[df.tmb$age > 0])
}
\arguments{
\item{df.tmb}{list of tmb parameters}

\item{reps}{estimated model.}

\item{Fbarage}{Fbar age}
}
\value{

}
\description{
Title
}
\examples{
smsplots(df.tmb, reps)

}
