% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smsPlots.R
\name{smsPlots}
\alias{smsPlots}
\title{Plot standard stock assessment figures}
\usage{
smsPlots(
  df.tmb,
  sas,
  type = "default",
  Blim = NULL,
  Bpa = NULL,
  printFig = TRUE
)
}
\arguments{
\item{df.tmb}{list of tmb parameters}

\item{sas}{estimated model.}

\item{type}{default, stack, or wrap}

\item{Blim}{optional value to graph if type is stack or wrap}

\item{Bpa}{optional value to add Bpa (Bescape)}

\item{printFig}{TRUE FALSE to return figure as list}
}
\value{
Returns a figure with SSB, Fbar, recruitment and catch
}
\description{
Plot standard stock assessment figures
}
\examples{
smsPlots(df.tmb, sas)
smsPlots(df.tmb, sas, type = "stack", Blim = 110000)
smsPlots(df.tmb, sas, type = "wrap", Blim = 110000)
}
