% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smsR-package.R
\docType{package}
\name{smsR-package}
\alias{smsR}
\alias{smsR-package}
\title{smsR: Stochastic MultiSpecies stock assessment model}
\description{
smsR is an R implementation of the SMS model used to run seasonal stock assessments for exploited marine populations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/yourusername/smsR # Replace with your actual URL}
}

}
\author{
\strong{Maintainer}: Anonymous Anonymous \email{Anonymous@secret.com} (\href{https://orcid.org/Anonymous}{ORCID})

}
\keyword{internal}
