% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smsR-package.R
\docType{package}
\name{smsR-package}
\alias{smsR}
\alias{smsR-package}
\title{smsR: Stochastic MultiSpecies stock assessment model}
\description{
smsR is an R implementation of the SMS model used to run seasonal stock assessments for exploited marine populations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nissandjac/smsR}
}

}
\author{
\strong{Maintainer}: Nis Sand Jacobsen \email{nsja@aqua.dtu.dk} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
