% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_to_matrix.R
\name{survey_to_matrix}
\alias{survey_to_matrix}
\title{Title}
\usage{
survey_to_matrix(survey, year, season = 1:4)
}
\arguments{
\item{survey}{data frame containing survey}

\item{year}{years  to plot}

\item{season}{which seasons  are modeled}
}
\value{

}
\description{
Title
}
\examples{
survey_to_matrix(x,y, season = 1:2)
}
