% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_to_matrix.R
\name{survey_to_matrix}
\alias{survey_to_matrix}
\title{Take a dataframe of surveys and converts them to matrix for smsR}
\usage{
survey_to_matrix(survey, year)
}
\arguments{
\item{survey}{data frame containing survey}

\item{year}{years  to plot}

\item{season}{which seasons  are modeled}
}
\value{
returns a matrix of observed surveys
}
\description{
Take a dataframe of surveys and converts them to matrix for smsR
}
\examples{
survey_to_matrix(x, y)
}
